+++ Specs/Scene/MaterialSpec.js
@@ -237,6 +237,10 @@ defineSuite([
         verifyPolylineMaterial('PolylineArrow');
     });
 
+    it('draws PolylineGlow built-in material', function() {
+        verifyPolylineMaterial('PolylineGlow');
+    });
+
     it('draws PolylineOutline built-in material', function() {
         verifyPolylineMaterial('PolylineOutline');
     });